pyflask/day10/flaskrestful02/requirements.txt
flask
flask-restful
pymysql
flask-sqlalchemy
flask-script
flask-migrate
